- backlightThreshold is now 0-100 instead 0-1 - better performance for brightness - use piecewise linear instead of sqrt

+++ include/utils/RgbTransform.h
@@ -46,10 +46,10 @@ public:
 	void setGamma(double gammaR,double gammaG=-1, double gammaB=-1);
 
 	/// @return The current lower brightness
-	double getBacklightThreshold() const;
+	int getBacklightThreshold() const;
 
 	/// @param backlightThreshold New lower brightness
-	void setBacklightThreshold(double backlightThreshold);
+	void setBacklightThreshold(int backlightThreshold);
 
 	/// @return The current state
 	bool getBacklightColored() const;
