+++ src/hyperion-remote/JsonConnection.h
@@ -154,7 +154,7 @@ public:
 		double *gammaR,
 		double *gammaG,
 		double *gammaB,
-		double *backlightThreshold,
+		int    *backlightThreshold,
 		int    *backlightColored,
 		int    *brightness,
 		int    *brightnessC);
